Move language object sorting and comparison utilities into core i18n module.
